requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
distlib==0.3.6
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
fire==0.5.0
lxml==4.9.2
pdf2docx==0.5.1
platformdirs==2.5.2
PyJWT==2.6.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
python-docx==0.8.11
pytz==2022.7.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
termcolor==2.2.0
tzdata==2022.7
virtualenv==20.16.5
